Moved angularCancelOnNavigate module in core package.
(cherry picked from commit 2a3e89b)
